Move api functions into multiqc/core/interactive.py
